add debug file , norm

+++ libftex/ft_split_memory_cleaner.c
@@ -6,7 +6,7 @@
 /*   By: kousuzuk <kousuzuk@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/09 16:50:28 by kousuzuk          #+#    #+#             */
-/*   Updated: 2023/10/09 16:53:01 by kousuzuk         ###   ########.fr       */
+/*   Updated: 2023/10/09 17:58:30 by kousuzuk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_split_memory_cleaner(size_t memory_len, char **array)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < memory_len)
